More NumPy compatibility work
Most uses of the typecode() method were for creating an array of the
same type as that of a given array. The utility functions zeros_st()
(zeros of same type) was provided for this task in Scientific.N; its
implementation is different when NumPy is used. The only other use of
the typecode() method (in BSP/core.py) was handled by catching the
AttributeError at runtime.